+++ lib/config.js
@@ -53,4 +53,15 @@ export class SettingsFunc {
             return true;
         }
     }
+
+    //加载设置
+    async loadConfig() {
+        try {
+            const fileContents = await fs.readFile(eCfgPath, 'utf8');
+            return yaml.parse(fileContents) || {};
+        } catch (error) {
+            logger.error('[esca-plugin] 加载配置文件失败:', error);
+            throw new Error('无法读取配置文件，请检查路径或文件权限');
+        }
+    }
 }
